requirements.txt
djangorestframework
Django==2.2.8
pytz==2019.3
sqlparse==0.3.0
